:truck: move issuer connection tests to issuer module
